@@ -9,17 +9,10 @@ import pprint
 import traceback
 from io import open  # pylint: disable=redefined-builtin
 from io import StringIO
+from pathlib import Path
 
 from importlib import resources
 from importlib.metadata import entry_points
-
-# Remove when we drop support for 3.8
-if sys.version_info < (3, 9):
-    import importlib_resources as resources
-
-# Remove when we drop support for 3.9
-if sys.version_info < (3, 10):
-    from importlib_metadata import entry_points
 
 
 SIG_FILES = [os.path.join('demo', f) for f in (
@@ -64,7 +57,7 @@ def _runscript(scriptname):
     if os.path.isfile(scriptfile):
         if os.path.isfile(scriptfile):
             exec(  # pylint: disable=exec-used
-                compile(open(scriptfile).read(), scriptfile, 'exec'),
+                compile(Path(scriptfile).read_text(), scriptfile, 'exec'),
                 namespace)
             return 0
 
